requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
passlib[bcrypt] 
python-jose[cryptography]
pydantic-settings
python-dotenv
psycopg2-binary
pydantic[email]
python-multipart
python-docx
